+++ react/inputs/StyleguideInput/index.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 import { intlShape, injectIntl } from 'react-intl'
 import Input from '@vtex/styleguide/lib/Input'
@@ -7,16 +7,28 @@ import ProfileFieldShape from '../../ProfileFieldShape'
 import GenderInput from './GenderInput'
 import styles from '../../styles.css'
 
-const StyleguideInput = props => {
-  const {
-    field,
-    data,
-    options,
-    inputRef,
-    onChange,
-    onBlur,
-    intl,
-  } = props
+const StyleguideInput = (props) => {
+  const { field, data, options, inputRef, onChange, onBlur, intl } = props
+
+  const [isFocused, setIsFocused] = useState(false)
+
+  // Handle focus and blur events
+  const handleFocus = () => setIsFocused(true)
+  const handleBlur = (e) => {
+    onBlur(e)
+    setIsFocused(false)
+  }
+
+  // Define class names based on the component state and props
+  const containerClassNames = [
+    styles.styleguideInput,
+    field.hidden ? 'dn' : '',
+    field.required ? `${styles.styleguideInput}-required` : '',
+    isFocused ? `${styles.styleguideInput}-focused` : '',
+    data.error ? `${styles.styleguideInput}-invalid` : '',
+    !data.value ? `${styles.styleguideInput}-empty` : '',
+    'pb7',
+  ].join(' ')
 
   if (field.name === 'gender') {
     return (
@@ -28,11 +40,7 @@ const StyleguideInput = props => {
   }
 
   return (
-    <div
-      className={`${styles.styleguideInput} ${
-        field.hidden ? 'dn' : ''
-      } pb7`}
-    >
+    <div className={containerClassNames}>
       <Input
         name={field.name}
         label={intl.formatMessage({
@@ -51,7 +59,8 @@ const StyleguideInput = props => {
             : null
         }
         onChange={onChange}
-        onBlur={onBlur}
+        onBlur={handleBlur}
+        onFocus={handleFocus}
         ref={inputRef}
         maxLength={field.maxLength}
         disabled={field.disabled}
@@ -75,6 +84,11 @@ StyleguideInput.propTypes = {
   onBlur: PropTypes.func.isRequired,
   /** React-intl utility */
   intl: intlShape.isRequired,
+}
+
+StyleguideInput.defaultProps = {
+  options: {},
+  inputRef: () => {},
 }
 
 export default injectIntl(StyleguideInput)