+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- "...-required", "...-disabled", "...-empty", "...-focused" and "...-invalid" classNames for the StyleguideInput component.
+
 ## [3.18.0] - 2024-09-18
 
 ### Added